Suggested fixes.
