+++ command/server/listener_test.go
@@ -62,14 +62,3 @@ func testListenerImpl(t *testing.T, ln net.Listener, connFn testListenerConnFn, 
 		t.Fatalf("bad: %v", buf.String())
 	}
 }
-
-func TestParseCiphers(t *testing.T) {
-	testOk := "TLS_ECDHE_ECDSA_WITH_AES_128_CBC_SHA:TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256:TLS_ECDHE_ECDSA_WITH_AES_256_CBC_SHA:TLS_ECDHE_ECDSA_WITH_AES_256_GCM_SHA384:TLS_ECDHE_RSA_WITH_AES_128_CBC_SHA:TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256:TLS_ECDHE_RSA_WITH_AES_256_CBC_SHA:TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384:TLS_RSA_WITH_AES_128_CBC_SHA:TLS_RSA_WITH_AES_128_GCM_SHA256:TLS_RSA_WITH_AES_256_CBC_SHA:TLS_RSA_WITH_AES_256_GCM_SHA384"
-	if _, err := parseCiphers(testOk); err != nil {
-		t.Fatal(err)
-	}
-	testBad := "cipher1:cipher2"
-	if _, err := parseCiphers(testBad); err == nil {
-		t.Fatal("should fail")
-	}
-}
